refactor: clean code

+++ src/K_pi/Data/Diff.php
@@ -6,13 +6,16 @@ namespace K_pi\Data;
 
 final class Diff
 {
-    public readonly int|float $diff;
+    public readonly float|int $diff;
 
     public readonly bool $changed;
 
-    public function __construct(public readonly string $name, public readonly int|float $from, public readonly int|float $to)
-    {
-        $this->diff = $to - $from;
+    public function __construct(
+        public readonly string $name,
+        public readonly float|int $from,
+        public readonly float|int $to,
+    ) {
+        $this->diff    = $to - $from;
         $this->changed = 0 != $this->diff;
     }
 }
